3 A/requirements.txt
amqp==5.1.1
asgiref==3.7.2
attrs==23.1.0
billiard==4.1.0
boto3==1.28.66
botocore==1.31.66
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
Django==4.2.5
django-celery-beat==2.5.0
django-ckeditor==6.7.0
django-cors-headers==4.3.0
django-js-asset==2.1.0
django-storages==1.14.2
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.4.2
drf-spectacular==0.26.5
eventlet==0.33.3
gevent==23.9.1
greenlet==3.0.0
idna==3.4
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kavenegar==1.1.2
kombu==5.3.2
Pillow==10.1.0
prompt-toolkit==3.0.39
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.8
zope.event==5.0
zope.interface==6.1
